% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-vs-dataSet.R
\name{EvaluateBoard}
\alias{EvaluateBoard}
\title{Implements the minimax algorithm}
\usage{
EvaluateBoard(localBoard)
}
\arguments{
\item{board:}{vector: the game board.}
}
\value{
an integer:
  0:  no one has one
  1:  player 1 has won
  -1: player -1 has won
}
\description{
Implements the minimax algorithm
}
\examples{
EvaluateBoard(rep(1:9))

}
