% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-vs-dataSet.R
\name{InitilizeGameTreeConversion}
\alias{InitilizeGameTreeConversion}
\title{The data.tree representing all games in tic-tac-toe, takes advantage of rotating
the board 90 degrees is the exact same board.  Based on initMove this function sets the translation
array: gameTreeConversionMapping.}
\usage{
InitilizeGameTreeConversion(initMove)
}
\arguments{
\item{initMove}{An integer.  The human's first click on the board.}
}
\value{
a function that takes two parameters:
    conversion: String ['toTree', 'toGame']
    value: Integer The value to be converted.
}
\description{
The data.tree representing all games in tic-tac-toe, takes advantage of rotating
the board 90 degrees is the exact same board.  Based on initMove this function sets the translation
array: gameTreeConversionMapping.
}
\examples{
InitilizeGameTreeConversion(5)

}
