% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-vs-dataSet.R
\name{MiniMax}
\alias{MiniMax}
\title{Implements the minimax algorithm}
\usage{
MiniMax(board, player, tree)
}
\arguments{
\item{board:}{vector: the game board.}

\item{player:}{integer: The player who's move it is.}

\item{player:}{integer: The player who's move it is.}
}
\value{
a list:
  childValue: integer: The chosen value for the children.
  child: child of tree: corresponds to bestValu
}
\description{
Implements the minimax algorithm
}
\examples{
MiniMax(rep(1:9), 1, tree)

}
