% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-vs-human.R, R/human-vs-tree.R
\name{server}
\alias{server}
\alias{server}
\title{figures out which was the first pick in the UI and that is the humanSide.}
\usage{
server(input, output, session)

server(input, output, session)
}
\arguments{
\item{board:}{3x3 matrix}

\item{index:}{Integer (1:9)}

\item{winner:}{Integer: who won the current game
localBoard: 3x3 matrix}

\item{input:}{shiny input}

\item{output:}{shiny output}

\item{session:}{shiny session}

\item{tree:}{tree representing the remaining possible moves for the game}

\item{localBoard:}{Vector 1:9}

\item{side:}{Integer: 1,-1}

\item{r:}{Integer: 1:3}

\item{c:}{Integer: 1:3}

\item{winner:}{Integer: 1,-1  Who won the name}

\item{localBoard:}{Vector 1:9}

\item{computerSide:}{Integer: 1,-1}

\item{tree:}{tree representing the remaining possible moves for the game}

\item{winner:}{Integer: 1,-1  Who won the name}

\item{localBoard:}{Vector 1:9}

\item{localBoard:}{Vector 1:9}

\item{tree:}{represting the last move by computer before human move}
}
\value{
Integer: 1,-1,0

list of UI checkbox names.

: Boolean: whether game is over not

List: localBoard, move, childTree

Integer  1:9 location in localBoard.

List  localBoard, move, tree for move

Boolean gameOver

child tree
}
\description{
figures out which was the first pick in the UI and that is the humanSide.

shiny server
}
\examples{


Resets for new game.

Converts a board location to a UI cell location.



called if New Game button is called.


If game over then update xWins, OWins, & ties.  It
also updates the UI.




reset game


convert row and column to a location in the localBoard.


convert row and column to a location in the localBoard.


compute computer move.


Update wins by X and O.  Also, show ties.


Find the child node of the current tree that
the human has made.


}
