% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-vs-dataSet.R
\name{tic.tac.toe}
\alias{tic.tac.toe}
\title{Main game}
\usage{
tic.tac.toe(ttt, player1 = "human", player2 = "computer")
}
\arguments{
\item{ttt:}{data.tree: the full list of moves for every game.}

\item{player1:}{String: the optional name of player1.}

\item{player2:}{String: the optional name of player2.}

\item{game:}{vector of length 9}

\item{ttt:}{data.tree: the full list of moves for every game.}

\item{player1:}{String: the optional name of player1.}

\item{player2:}{String: the optional name of player2.}

\item{node:}{data.tree: the full list of moves for every game.}

\item{label:}{String: the optional description of which node this represents}
}
\value{
an integer:
  0:  no one has one
  1:  player 1 has won
  -1: player -1 has won

an integer:
  0:  no one has one
  1:  player 1 has won
  -1: player -1 has won
}
\description{
Main game
}
\examples{
ttt <- TTTTree()
tic.tac.toe(ttt)

Draw the game board - using plots

ttt <- TTTTree()
draw.board(rep(1:9))

Main game

ttt <- TTTTree()
tic.tac.toe(ttt)

print a node in the tree.

PrintNode(node)
tic.tac.toe(ttt, label='child')

}
